camp-python-2021-cofounders-finder-develop/requirements/production.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/production.in
#
aioredis==1.3.1
    # via channels-redis
amqp==5.0.6
    # via kombu
asgiref==3.4.1
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
async-timeout==3.0.1
    # via aioredis
attrs==21.2.0
    # via
    #   automat
    #   jsonschema
    #   pytest
    #   service-identity
    #   twisted
autobahn==21.3.1
    # via daphne
automat==20.2.0
    # via twisted
billiard==3.6.4.0
    # via celery
boto3==1.18.3
    # via django-storages
botocore==1.21.3
    # via
    #   boto3
    #   s3transfer
celery[redis]==5.1.2
    # via
    #   -r requirements/production.in
    #   celery-haystack-ng
    #   django-celery-beat
celery-haystack-ng==0.20.post2
    # via -r requirements/production.in
cffi==1.14.6
    # via cryptography
channels==3.0.4
    # via
    #   -r requirements/production.in
    #   channels-redis
channels-redis==3.3.0
    # via -r requirements/production.in
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
colorama==0.4.4
    # via rich
commonmark==0.9.1
    # via rich
constantly==15.1.0
    # via twisted
cryptography==3.4.7
    # via
    #   autobahn
    #   django-rest-knox
    #   pyopenssl
    #   service-identity
daphne==3.0.2
    # via channels
django==3.2.5
    # via
    #   -r requirements/production.in
    #   channels
    #   django-appconf
    #   django-cacheops
    #   django-celery-beat
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-haystack
    #   django-health-check
    #   django-redis
    #   django-rest-knox
    #   django-storages
    #   django-timezone-field
    #   djangorestframework
    #   drf-haystack
    #   drf-spectacular
django-appconf==1.0.4
    # via
    #   celery-haystack-ng
    #   django-imagekit
django-cacheops==6.0
    # via -r requirements/production.in
django-celery-beat==2.2.1
    # via -r requirements/production.in
django-cors-headers==3.7.0
    # via -r requirements/production.in
django-extensions==3.1.3
    # via -r requirements/production.in
django-filter==2.4.0
    # via -r requirements/production.in
django-haystack==3.0.0
    # via
    #   -r requirements/production.in
    #   celery-haystack-ng
    #   drf-haystack
django-health-check==3.16.4
    # via -r requirements/production.in
django-imagekit==4.0.2
    # via -r requirements/production.in
django-redis==5.0.0
    # via -r requirements/production.in
django-rest-knox==4.1.0
    # via -r requirements/production.in
django-storages[boto3]==1.11.1
    # via -r requirements/production.in
django-timezone-field==4.2.1
    # via django-celery-beat
djangorestframework==3.12.4
    # via
    #   -r requirements/production.in
    #   django-rest-knox
    #   drf-haystack
    #   drf-spectacular
drf-haystack==1.8.10
    # via -r requirements/production.in
drf-spectacular==0.17.2
    # via -r requirements/production.in
execnet==1.9.0
    # via pytest-xdist
factory-boy==3.2.0
    # via -r requirements/production.in
faker==8.10.1
    # via factory-boy
funcy==1.16
    # via django-cacheops
hiredis==2.0.0
    # via aioredis
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.2
    # via
    #   hyperlink
    #   twisted
incremental==21.3.0
    # via twisted
inflection==0.5.1
    # via drf-spectacular
iniconfig==1.1.1
    # via pytest
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonschema==3.2.0
    # via drf-spectacular
kombu==5.1.0
    # via celery
mistune==0.8.4
    # via -r requirements/production.in
msgpack==1.0.2
    # via channels-redis
packaging==21.0
    # via
    #   pytest
    #   pytest-sugar
pilkit==2.0
    # via django-imagekit
pillow==8.3.1
    # via -r requirements/production.in
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.19
    # via click-repl
psycopg2-binary==2.9.1
    # via -r requirements/production.in
py==1.10.0
    # via
    #   pytest
    #   pytest-forked
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycparser==2.20
    # via cffi
pygments==2.9.0
    # via rich
pyopenssl==20.0.1
    # via twisted
pyparsing==2.4.7
    # via packaging
pyrsistent==0.18.0
    # via jsonschema
pytest==6.2.4
    # via
    #   -r requirements/production.in
    #   pytest-django
    #   pytest-forked
    #   pytest-sugar
    #   pytest-xdist
pytest-django==4.4.0
    # via -r requirements/production.in
pytest-forked==1.3.0
    # via pytest-xdist
pytest-sugar==0.9.4
    # via -r requirements/production.in
pytest-xdist==2.3.0
    # via -r requirements/production.in
python-crontab==2.5.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   botocore
    #   drf-haystack
    #   faker
    #   python-crontab
pytz==2021.1
    # via
    #   celery
    #   django
    #   django-timezone-field
pyuwsgi==2.0.19.1.post0
    # via -r requirements/production.in
pyyaml==5.4.1
    # via drf-spectacular
redis==3.5.3
    # via
    #   celery
    #   django-cacheops
    #   django-redis
rich==10.6.0
    # via -r requirements/production.in
s3transfer==0.5.0
    # via boto3
service-identity==21.1.0
    # via twisted
six==1.16.0
    # via
    #   automat
    #   click-repl
    #   django-cacheops
    #   django-imagekit
    #   jsonschema
    #   pyopenssl
    #   python-dateutil
    #   service-identity
sqlparse==0.4.1
    # via django
termcolor==1.1.0
    # via pytest-sugar
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via pytest
twisted[tls]==21.7.0
    # via daphne
txaio==21.2.1
    # via autobahn
typing-extensions==3.10.0.0
    # via twisted
uritemplate==3.0.1
    # via drf-spectacular
urllib3==1.26.6
    # via botocore
uwsgitop==0.11
    # via -r requirements/production.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
zope.interface==5.4.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
